+++ src/contentful/index.ts
@@ -1,2 +1,4 @@
+import { getContentfulRequest } from './utils';
+
 export * from './utils';
 export * from './schema';