+++ src/index.ts
@@ -1,8 +1,21 @@
-import * as analytics from './analytics';
+export {
+  default as createScrollManager,
+} from './scrollmanager';
+export {
+  configureAnalytics,
+  createAnalytics,
+  track,
+  getUTMKeysFromSession,
+  setGTMLoaded,
+  setEnabled as setAnalyticsEnabled,
+} from './analytics';
 
-export { analytics };
-export * from './contentful';
-export * from './scrollmanager';
-
-
-// TODO: fix all those BS exports
+export {
+  configureContentful,
+  createContentfulRequest,
+  getContentfulRequest,
+  getSchema as getContentfulSchema,
+  getSchemaOptions as getContentfulSchemaOptions,
+  formatImage as formatContentfulImage,
+  formatImages as formatContentfulImages,
+} from './contentful';